move new tests into original test file
